Refactor box-width helpers per review
- Move text_width to utils (it is a general measurement helper, not
  config), and name its magic numbers: POINTS_PER_INCH and the existing
  TEXT_WIDTH_RESOLUTION.
- Fold the drawing.py module-level helpers _draws_label and
  _box_min_width into DRAWING_ATTRIBUTES as the "draws_label" and
  "box_min_width" attributes, computed once per box like the others;
  drawing.py now reads box.box_min_width directly.
- Drop box_label_width's drawing_name/name fallback: drawing_name is
  already guaranteed set by the "drawing_name" attribute that precedes
  it in DRAWING_ATTRIBUTES.

Pure refactor: baseline images and tikz are unchanged.
